Modifying MH tests to be merged with the Docker tests, instead of being in a separate directory
