--- test/docker-test/omsfiles/customlog.conf
@@ -1,11 +1,0 @@
-# This file is configured by the OMS service
-
-<source>
-  type sudo_tail
-  path /var/log/custom.log
-  pos_file /var/opt/microsoft/omsagent/state/CUSTOM_LOG_BLOB.Custom_Log_CL_<workspace-id>.pos
-  read_from_head false
-  run_interval 60
-  tag oms.blob.CustomLog.CUSTOM_LOG_BLOB.Custom_Log_CL_<workspace-id>.*
-  format none
-</source>
