+++ test/docker-test/omsfiles/workspace_1/customlog.conf
