+++ test/docker-test/omsfiles/workspace_1/rsyslog-oms.conf
