+++ test/docker-test/omsfiles/workspace_2/customlog.conf
