+++ test/docker-test/omsfiles/workspace_2/rsyslog-oms.conf
