--- test/multihoming-test/omsfiles/workspace_1/rsyslog-oms.conf
@@ -1,6 +1,0 @@
-
-# OMS Syslog collection for workspace a0d166ba-98b9-402e-b805-172ed62150a4
-daemon.=alert;daemon.=crit;daemon.=debug;daemon.=emerg;daemon.=err;daemon.=info;daemon.=notice;daemon.=warning	@127.0.0.1:25224
-kern.=alert;kern.=crit;kern.=debug;kern.=emerg;kern.=err;kern.=info;kern.=notice;kern.=warning	@127.0.0.1:25224
-syslog.=alert;syslog.=crit;syslog.=debug;syslog.=emerg;syslog.=err;syslog.=info;syslog.=notice;syslog.=warning	@127.0.0.1:25224
-cron.=alert;cron.=crit;cron.=debug;cron.=emerg;cron.=err;cron.=info;cron.=notice;cron.=warning	@127.0.0.1:25224
