--- test/multihoming-test/omsfiles/workspace_2/mysql-slow.log
@@ -1,11 +1,0 @@
-# Time: 2018-10-15T19:56:12.584806Z
-# User@Host: root[root] @ localhost []  Id:     8
-# Query_time: 10.000575  Lock_time: 0.000000 Rows_sent: 1  Rows_examined: 0
-use test;
-SET timestamp=1539633372;
-select sleep(10);
-# Time: 2018-10-15T19:56:36.398796Z
-# User@Host: root[root] @ localhost []  Id:     8
-# Query_time: 10.000546  Lock_time: 0.000000 Rows_sent: 1  Rows_examined: 0
-SET timestamp=1539633396;
-select sleep(10);
