--- test/multihoming-test/omsfiles/workspace_2/perf.conf
@@ -1,36 +1,0 @@
-
-<source>
-  type oms_omi
-  object_name "Logical Disk"
-  instance_regex ".*"
-  counter_name_regex "(% Used Inodes|Free Megabytes|% Used Space|Disk Transfers/sec|Disk Reads/sec|Disk Writes/sec)"
-  interval 10s
-  omi_mapping_path /etc/opt/microsoft/omsagent/<workspace-id>/conf/omsagent.d/omi_mapping.json
-</source>
-
-<source>
-  type oms_omi
-  object_name "Processor"
-  instance_regex ".*"
-  counter_name_regex "(% Processor Time|% Privileged Time)"
-  interval 10s
-  omi_mapping_path /etc/opt/microsoft/omsagent/<workspace-id>/conf/omsagent.d/omi_mapping.json
-</source>
-
-<source>
-  type oms_omi
-  object_name "Network"
-  instance_regex ".*"
-  counter_name_regex "(Total Bytes Transmitted|Total Bytes Received)"
-  interval 10s
-  omi_mapping_path /etc/opt/microsoft/omsagent/<workspace-id>/conf/omsagent.d/omi_mapping.json
-</source>
-
-<source>
-  type oms_omi
-  object_name "Memory"
-  instance_regex ".*"
-  counter_name_regex "(Available MBytes Memory|% Used Memory|% Used Swap Space)"
-  interval 10s
-  omi_mapping_path /etc/opt/microsoft/omsagent/<workspace-id>/conf/omsagent.d/omi_mapping.json
-</source>
