--- test/multihoming-test/parameters.json
@@ -1,16 +1,0 @@
-{
-  "resource": "https://management.azure.com",
-  "tenant": "72f988bf-86f1-41af-91ab-2d7cd011db47",
-  "authority host url": "https://login.microsoftonline.com",
-  "app id": "<app-id>",
-  "app secret": "<app-secret>",
-  "oms bundle": "<bundle-file-name>",
-  "resource group": "<resource-group-name>",
-  "subscription": "<subscription-id>",
-  "workspace 1": "<workspace-1-name>",
-  "workspace 2": "<workspace-2-name>",
-  "workspace id 1": "<workspace-1-id>",
-  "workspace key 1": "<workspace-1-key>",
-  "workspace id 2": "<workspace-2-id>",
-  "workspace key 2": "<workspace-2-key>"
-}
